synthetic_datagen_project/requirements.txt
anthropic>=0.25.0
pyyaml>=6.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
mem0ai>=0.1.0
